Fontify based on rendered column widths
When tabulated-list-mode renders a line, it considers the remaining
available space in the selected window and may reduce the width of a
column compared to the width specified in tabulated-list-format.  As a
result, fontifying or otherwise calculating offsets based on the
widths in tabulated-list-format can lead to colors bleeding over into
the next column.

Delete function pocket-reader--column-data, which returns the
"theoretical" start and end positions of each column, and instead use
the actual text properties in the buffer.

Fixes #36.