Fix: (pocket-reader--get-items) Use JSON key for article list
Somehow, for years (ever since this library's inception), the JSON
objects had the keys in a consistent order, so getting the third item
worked (though I still don't know why I would have written it that
way).  Even now, as I test it, the list of items appears to always be
the third element.  But for some users, this is apparently not the
case sometimes.  I don't know if it's caused by the server or by a
JSON parsing library, etc.

Anyway, we now get the list by key, which is correct.

Fixes #53.

Reported-by: Christof Damian <https://github.com/christofdamian>
Reported-by: radixm46 <https://github.com/radixm46>